+++ src/bin/commands/index.ts
@@ -1,3 +1,4 @@
 import { create } from './create';
+import { update } from './update';
 
-export const commands = [create];
+export const commands = [create, update];