+++ src/utils/merge-template.ts
@@ -6,8 +6,13 @@ import { baseUrl, Template, templates } from '../templates';
  *
  * @param runner Command runner
  * @param type Template type
+ * @param isExistingProject (Optional) Pass true for existing projects
  */
-export function mergeTemplate(runner: CommandRunner, type: Template): void {
+export function mergeTemplate(
+	runner: CommandRunner,
+	type: Template,
+	isExistingProject?: boolean
+): void {
 	const remote = 'template-' + type;
 	const url = templates[type] ?? baseUrl + type;
 
@@ -24,9 +29,16 @@ export function mergeTemplate(runner: CommandRunner, type: Template): void {
 		loadingDescription: 'Fetching',
 	});
 
-	runner.run(`git pull ${remote} master --allow-unrelated`, {
-		loadingDescription: 'Pulling',
-	});
+	if (isExistingProject) {
+		runner.run(`git merge ${remote}/master --allow-unrelated`, {
+			loadingDescription: 'Merging',
+		});
+	}
+	else {
+		runner.run(`git pull ${remote} master --allow-unrelated`, {
+			loadingDescription: 'Pulling',
+		});
+	}
 
 	runner.run(`git remote remove ${remote}`, {
 		loadingDescription: 'Cleaning up',