+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## [1.0.0] Initial Release
 
 ## Additions
+- [Issue #2] As a user, I want to update an existing project
 - [Issue #1] As a user, I want to create a new project